Release v4.8.7.0
### CLI

- Fixed accidental ignoring of CLI arguments: `-v`, `-ver`, `-version`
- Improved CLI help formatting
- Improved build script

### CSScriptLib

- no changes

+++ src/3.build-documentation.cmd
@@ -1,17 +1,16 @@
 echo off
 
-.\src\out\Windows\css.exe -server_r:stop
-.\src\out\Windows\css.exe -server:stop
+.\out\Windows\css.exe -server_r:stop
+.\out\Windows\css.exe -server:stop
 
-set vs_edition=Community
-if exist "C:\Program Files\Microsoft Visual Studio\2022\%vs_edition%" (
-    echo Visual Studio 2022 (Community)
-) else (
-    set vs_edition=Professional
-    echo Visual Studio 2022 (PRO)
-)
+set vs_edition=undefined
+if exist "C:\Program Files\Microsoft Visual Studio\2022\Community" set vs_edition=Community
+if exist "C:\Program Files\Microsoft Visual Studio\2022\Professional" set vs_edition=Professional
 
-rd /Q /S .\CSScriptLib.Doc\help
+echo Visual Studio 2022 (%vs_edition%)
+
+if exist .\CSScriptLib.Doc\help rd /Q /S .\CSScriptLib.Doc\help
+
 set msbuild="C:\Program Files\Microsoft Visual Studio\2022\%vs_edition%\MSBuild\Current\Bin\MSBuild.exe"
 %msbuild% ".\CSScriptLib.Doc\CSScriptLib.Doc.shfbproj" -p:Configuration=Release -t:rebuild /p:WarningLevel=0
 
@@ -21,3 +20,4 @@ cd ".\out\Windows"
 cd ..\..\
 
 explorer .\out
+
